requirements.txt
numpy>=1.13
scipy>=0.19
pandas>=0.20
scikit-learn>=0.19
matplotlib>=2.0.2
tensorflow>=1.3
keras>=2.0.8
google-apputils>=0.4.2

